# Optimize log injection with Serilog