--- a/test/Datadog.Trace.Tests/Logging/SerilogLogProviderTests.cs
+++ b/test/Datadog.Trace.Tests/Logging/SerilogLogProviderTests.cs
@@ -1,7 +1,8 @@
 using System;
 using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 using Datadog.Trace.Logging;
-using Datadog.Trace.Logging.LogProviders;
 using Serilog;
 using Serilog.Events;
 using Xunit;
@@ -23,7 +24,7 @@ namespace Datadog.Trace.Tests.Logging
                 .CreateLogger();
             _logEvents = new List<LogEvent>();
 
-            _logProvider = new SerilogLogProvider();
+            _logProvider = new CustomSerilogLogProvider();
             LogProvider.SetCurrentLogProvider(_logProvider);
             _logger = new LoggerExecutionWrapper(_logProvider.GetLogger("Test"));
         }
@@ -32,7 +33,7 @@ namespace Datadog.Trace.Tests.Logging
         public void LogsInjectionEnabledAddsParentCorrelationIdentifiers()
         {
             // Assert that the Serilog log provider is correctly being used
-            Assert.IsType<SerilogLogProvider>(LogProvider.CurrentLogProvider);
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
 
             // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
             var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: true);
@@ -47,7 +48,7 @@ namespace Datadog.Trace.Tests.Logging
         public void LogsInjectionEnabledAddsChildCorrelationIdentifiers()
         {
             // Assert that the Serilog log provider is correctly being used
-            Assert.IsType<SerilogLogProvider>(LogProvider.CurrentLogProvider);
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
 
             // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
             var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: true);
@@ -62,7 +63,7 @@ namespace Datadog.Trace.Tests.Logging
         public void LogsInjectionEnabledDoesNotAddCorrelationIdentifiersOutsideSpans()
         {
             // Assert that the Serilog log provider is correctly being used
-            Assert.IsType<SerilogLogProvider>(LogProvider.CurrentLogProvider);
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
 
             // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
             var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: true);
@@ -77,7 +78,7 @@ namespace Datadog.Trace.Tests.Logging
         public void LogsInjectionEnabledUsesTracerServiceName()
         {
             // Assert that the Serilog log provider is correctly being used
-            Assert.IsType<SerilogLogProvider>(LogProvider.CurrentLogProvider);
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
 
             // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
             var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: true);
@@ -92,15 +93,65 @@ namespace Datadog.Trace.Tests.Logging
         public void DisabledLibLogSubscriberDoesNotAddCorrelationIdentifiers()
         {
             // Assert that the Serilog log provider is correctly being used
-            Assert.IsType<SerilogLogProvider>(LogProvider.CurrentLogProvider);
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
 
-            // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
+            // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to FALSE
             var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: false);
             LoggingProviderTestHelpers.LogEverywhere(tracer, _logger, _logProvider.OpenMappedContext, out var parentScope, out var childScope);
 
             // Filter the logs
             _logEvents.RemoveAll(log => !log.MessageTemplate.ToString().Contains(LoggingProviderTestHelpers.LogPrefix));
             Assert.All(_logEvents, e => LogEventDoesNotContainCorrelationIdentifiers(e));
+        }
+
+        [Fact]
+        public void LogTheRightSpan()
+        {
+            // Assert that the Serilog log provider is correctly being used
+            Assert.IsType<CustomSerilogLogProvider>(LogProvider.CurrentLogProvider);
+
+            // Instantiate a tracer for this test with default settings and set LogsInjectionEnabled to TRUE
+            var tracer = LoggingProviderTestHelpers.InitializeTracer(enableLogsInjection: true);
+
+            var barrier1 = new Barrier(2);
+            var barrier2 = new Barrier(2);
+
+            var task1 = Task.Run(() => Thread1(barrier1, tracer));
+            barrier1.SignalAndWait(); // Wait until thread1 has created both scopes
+
+            var task2 = Task.Run(() => Thread2(barrier2, tracer));
+            barrier2.SignalAndWait(); // Wait until thread2 has created both scopes
+
+            // Stack is: Empty -> outer1 -> Empty -> outer2
+
+            barrier1.SignalAndWait(); // Release thread1
+            barrier1.SignalAndWait(); // Wait until thread1 has disposed the inner scope
+
+            // Stack is: Empty -> outer1 -> Empty
+
+            barrier2.SignalAndWait(); // Release thread2
+            // Thread2 closes the inner span and restores "Empty" from the stack: first log has no span even though it should
+            // Then thread2 closes the outer span and restores "outer1" from the stack: second log has a span even though it shouldn't
+            task2.Wait();
+
+            // Unblock the first thread
+            barrier1.SignalAndWait();
+            task1.Wait();
+
+            Assert.Equal(2, _logEvents.Count);
+
+            var spanLog = _logEvents[0];
+
+            Assert.StartsWith("Span", spanLog.MessageTemplate.Text);
+            Assert.Equal(5, spanLog.Properties.Count);
+
+            var expectedSpanId = spanLog.MessageTemplate.Text.Split('-')[1];
+            var spanProperty = spanLog.Properties[CorrelationIdentifier.SerilogSpanIdKey];
+            Assert.Equal(expectedSpanId, spanProperty.ToString().Trim('\"'));
+
+            var noSpanLog = _logEvents[1];
+            Assert.StartsWith("NoSpan", noSpanLog.MessageTemplate.ToString());
+            Assert.Empty(noSpanLog.Properties);
         }
 
         internal static void LogEventContains(Serilog.Events.LogEvent logEvent, string service, string version, string env, Scope scope)
@@ -132,6 +183,37 @@ namespace Datadog.Trace.Tests.Logging
             // Do not assert on the service property
             Assert.False(logEvent.Properties.ContainsKey(CorrelationIdentifier.SerilogSpanIdKey));
             Assert.False(logEvent.Properties.ContainsKey(CorrelationIdentifier.SerilogTraceIdKey));
+        }
+
+        private static void Thread1(Barrier barrier, Tracer tracer)
+        {
+            using (tracer.StartActive("Outer"))
+            {
+                using (tracer.StartActive("Inner"))
+                {
+                    barrier.SignalAndWait();
+                    barrier.SignalAndWait();
+                }
+
+                barrier.SignalAndWait();
+                barrier.SignalAndWait();
+            }
+        }
+
+        private static void Thread2(Barrier barrier, Tracer tracer)
+        {
+            using (var outerScope = tracer.StartActive("Outer"))
+            {
+                using (tracer.StartActive("Inner"))
+                {
+                    barrier.SignalAndWait();
+                    barrier.SignalAndWait();
+                }
+
+                Log.Information("Span-" + outerScope.Span.SpanId);
+            }
+
+            Log.Information("NoSpan");
         }
     }
 }