--- a/test/benchmarks/Benchmarks.Trace/Program.cs
+++ b/test/benchmarks/Benchmarks.Trace/Program.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Running;