--- a/test/benchmarks/Benchmarks.Trace/Properties/launchSettings.json
+++ b/test/benchmarks/Benchmarks.Trace/Properties/launchSettings.json
@@ -1,0 +1,19 @@
+{
+  "iisSettings": {
+    "windowsAuthentication": false,
+    "anonymousAuthentication": true,
+    "iisExpress": {
+      "applicationUrl": "http://localhost:56746/",
+      "sslPort": 44365
+    }
+  },
+  "profiles": {
+    "Benchmarks.Trace": {
+      "commandName": "Project",
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "https://localhost:5001;http://localhost:5000"
+    }
+  }
+}