--- a/test/benchmarks/Benchmarks.Trace/SerilogBenchmark.cs
+++ b/test/benchmarks/Benchmarks.Trace/SerilogBenchmark.cs
@@ -1,17 +1,18 @@
-using System.IO;
 using BenchmarkDotNet.Attributes;
 using Datadog.Trace;
 using Datadog.Trace.Configuration;
+using Datadog.Trace.Logging;
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
+using Serilog.Formatting.Display;
+using Logger = Serilog.Core.Logger;
 
 namespace Benchmarks.Trace
 {
     [MemoryDiagnoser]
     public class SerilogBenchmark
     {
-        private static readonly Logger EnrichedLogger;
         private static readonly Logger Logger;
         private static readonly Tracer LogInjectionTracer;
         private static readonly Tracer BaselineTracer;
@@ -38,16 +39,12 @@ namespace Benchmarks.Trace
             };
 
             BaselineTracer = new Tracer(baselineSettings, new DummyAgentWriter(), null, null, null);
-
-            EnrichedLogger = new LoggerConfiguration()
-                // Add Enrich.FromLogContext to emit Datadog properties
-                .Enrich.FromLogContext()
-                .WriteTo.Sink(new NullSink())
-                .CreateLogger();
+            var formatter = new MessageTemplateTextFormatter("{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] {Message:lj}{NewLine}{Exception}{Properties}{NewLine}", null);
 
             Logger = new LoggerConfiguration()
+                // Add Enrich.FromLogContext to emit Datadog properties
                 .Enrich.FromLogContext()
-                .WriteTo.Sink(new NullSink())
+                .WriteTo.Sink(new NullSink(formatter))
                 .CreateLogger();
         }
 
@@ -56,24 +53,57 @@ namespace Benchmarks.Trace
         {
             using (BaselineTracer.StartActive("Test"))
             {
-                Logger.Information("Hello");
+                using (BaselineTracer.StartActive("Child"))
+                {
+                    Logger.Information("Hello");
+                }
             }
         }
 
         [Benchmark]
         public void EnrichedLog()
         {
+            LibLogScopeEventSubscriber.UseOptimization = false;
+
             using (LogInjectionTracer.StartActive("Test"))
             {
-                EnrichedLogger.Information("Hello");
+                using (LogInjectionTracer.StartActive("Child"))
+                {
+                    Logger.Information("Hello");
+                }
+            }
+        }
+
+        [Benchmark]
+        public void EnrichedLogOptim()
+        {
+            LibLogScopeEventSubscriber.UseOptimization = true;
+
+            using (LogInjectionTracer.StartActive("Test"))
+            {
+                using (LogInjectionTracer.StartActive("Child"))
+                {
+                    Logger.Information("Hello");
+                }
             }
         }
 
         private class NullSink : ILogEventSink
         {
+            private readonly MessageTemplateTextFormatter _formatter;
+
+            public NullSink(MessageTemplateTextFormatter formatter)
+            {
+                _formatter = formatter;
+            }
+
             public void Emit(LogEvent logEvent)
             {
-                logEvent.RenderMessage(TextWriter.Null);
+#if DEBUG
+                _formatter.Format(logEvent, System.Console.Out);
+#else
+                _formatter.Format(logEvent, System.IO.TextWriter.Null);
+#endif
             }
         }
     }