- Use a single enricher to add properties to logs (LibLog uses `LogContext.PushProperty` which creates an enricher for _every_ property)
 - Cache the properties which value doesn't change over time (version, environment, service)
 - Convert `SpanEventArgs` to struct
 - Change the benchmark scenario to be more realistic: use nested spans and use a message template
 - Use AsyncLocal to flow the Serilog context instead of the shared stack. This fixes issues when span operations are interleaved between multiple threads. It also means we register the enricher only once per trace, instead of for every span.
 - Add a unit test for the previous point

Benchmark:
|           Method |     Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|              Log | 3.306 μs | 0.0440 μs | 0.0672 μs |  1.00 |    0.00 | 0.0153 |     - |     - |   1.48 KB |
|  EnrichedLog old | 6.934 μs | 0.0547 μs | 0.0457 μs |  2.07 |    0.05 | 0.0610 |     - |     - |   5.66 KB |
|  EnrichedLog new | 4.960 μs | 0.0585 μs | 0.0519 μs |  1.48 |    0.04 | 0.0305 |     - |     - |   2.67 KB |


There's still a significant overhead, but I'm afraid we can't do much more with the Serilog API. We'll see how it behaves on the reliability environment, but if we want to go further we should start thinking about using automatic instrumentation for log injection.